Permit bot class reloading
Add a ClassLoader for everything in the `bot` package which
effectively always reloads from disk. Erase the concrete
type of bots beind a new `BotProxy` class, which uses reflection
to actually invoke methods.

When a user reloads, we interrupt every live bot, clear the list
of bots, and then re-initialize the list from the classes loaded
through the ClassLoader.

Hopefully haven't run afoul of the object or memory models in
the particulars, but it seems to work as-is, and leaks are
not so much of a concern because this is just to make development
smoother.